refactor(gamesApi): initial relocation of questionCreatorFunc
